Convert Task modal to a Vue transitions modal.
See: https://vuejs.org/v2/examples/modal.html, https://vuejs.org/v2/guide/transitions.html
