+++ timestrap/static_src/sass/base.scss
@@ -6,4 +6,5 @@
 @import 'timer';
 @import 'entries';
 
+@import 'modal';
 @import 'select2';
